qos-gargoyle: Remove modification of a function parameter.
The parameter `cc` is not used after a brief check of its value, but we
decrease its value by `hlen`.

Since we are passing cc by value, the changes that we perform inside the
function don't get propagated to the caller, which means that it is totally
useless to make such computation.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>