Address uploading language pack by form. While it once worked, it seems to now be broken with a dropbear "Premature exit: Terminated by signal" error.
Bring the form method up to mirror how the internet-install method will work - by allowing firstboot to be in the native language as well. Sadly, that entails the use of an extra button to upload (thus getting around the dropbear issue), but allows getting rid of the timezone map, since timezones will be translated into the native language.