Update qos_gargoyle.init
ACC needs better automatic target than gateway, because when WAN is not PPPoE, probably  it is a ADSL/Cable modem. Using this kind of router is almost sure that ACC wont work properly.
This mod uses traceroute to find the next non private router in the next 4 hops, it is using -n -w 1 -q 2 to avoid lost to much time.