Update miniupnpd.init
From my testing, the source of Phgerin's problems is that after a sysupgrade (preserve settings) the firewall gets a little upset.
A simple restart of the firewall fixed these issues.
Steps to fix:
1. disable upnpd
2. issue firewall restart
3. reenable upnpd

so two proposed solutions, this fix here. 
OR
making sure that if we do a "preserve settings" the firewall gets properly flushed and restarted.

What do you think?
I have yet to test this change as i am still doing the 4GB transfer test.