Implement Edit & Remove Buttons
The Edit Button is done a little differently than std Gargoyle. In
std Gargoyle the Edit Button pops up and populates a form for editing.
Here, the Edited Row is Removed and the fields in the ADD form are
populated with the existing info so that the user can adjust and re-add.
This approach seems intuitive for the User and makes the JS code simpler
and smaller. If editing a really large table then the User may be
confused when the row dissapears while not being able to see the Add Row
at the top?