Integrate Device Groups into Quotas page
The Gagoyle admin can use a Device Group to define a Quota by selecting
"Only the following Host(s) and providing an IP, or IP range, or a Device
Group. The Group can be selected or entered manually and it is validated
an saved thru the existing processes. The Device Group MACs are included
in the validation to avoid quota overlaps involving MACs.

To integrate the MACs validation, the groupsValid parameter was added to
the function addAddressStringToTable() and this cascaded small
adjustments to other JavaScript files which called this function.