Fix guest access in Samba
The "valid users =" clause breaks the guest access.
By removing it, all expected behaviour is retained.
If the user wants write access, they need to map the drive with authentication credentials. Otherwise they will just be allowed to access the drive without a prompt with the permissions given to the guest (r/rw).

I have only commented the line out instead of removing it entirely to make the commit easier to revert in-case there is a usage case that i have not accounted for that does not behave correctly without the "valid users=" clause.