Added default_small profile
default_small is similar to default, however it disables debugging features
in kernel and strip kernel image and libraries to reduce the size. The result
is that small routers with 4MB of Flash will have more spare space, allowing
miniupnp package to be reactivated too (bringing UPnP support again, missing
since Gargoyle 1.9).

The consequence of stripping kernel image and libraries is that the resulting
firmware will not be compatible with packages from OpenWRT itself. Only the
packages compiled with the firmware will be compatible. However this should
not be a problem, all targets of this profile are routers with small Flash AND
no USB support. So it generally impossible to install too many things on it.

This commit reactivates miniupnp for default profile too, since the target
routers in this profile should be big enough to fit miniupnp.