Make Gargoyle theme files fallback to default files if they do not exist.
This enables themes to be compacted by only needing to provide files that they wish to edit.
/www/themes/THEME_NAME/theme.css is now a MANDATORY file, and is the recommended way that developers modify the
Gargoyle GUI. This file takes the place of common.css which used to be the mandatory file for themes.
For more advanced themes with more flexibility, by providing a new bootstrap.min.css, common.css or internal.css file
these will still override the default theme as expected.
If a developer wishes to override a default theme file, and does not wish to make use of that file at all, a blank
file with the same name can be provided to achieve this.

The above behaviour is consistent with all theme css and image files except for the madatory theme.css file.