Fix Green and Light theme when scrolling help.
In case of alignToTop true, #content jumps up behind the #sidebar-header of
fixed position due to margin-top 58px of #sidebar leaving free vertical space
for #content to float left. To solve this, #sidebar must begin at top 0. The
border is drawn by ul.sidebar instead. This also fixes the missing bottom border
of the sidebar compared to the bottom border of #content as it previously
overflowed the viewport bottom by 58px and thus no bottom border was drawn.

This fix is inspired by Matei theme.