Build all ath10k firmware & driver pkgs for ath79.[ath10k|ath10k-large]
While the ct (Candela Tech) driver and firmware packages are the
OpenWrt default, and should remain the Gargoyle default, some users
report better results with the original (aka non-ct) driver and
firmware packages.  As these are kmods, the OpenWrt packages can't
be used for Gargoyle so build them to give Gargoyle users the
option of installing them in place of the default ct packages.

Additionally there are now several variants of the ct firmware
packages and smallbuffers variants (which reduce memory
consumption) of the both ct and non-ct driver packages so build
them as well.