Ensure profile meta-packages are built
The profile meta-packages should be built by default but not forcibly
included in any images.

In the supported targets, the config device packages option is used
to specify the most appropriate package for any device.

When using make custom, the user will can select one meta-package to
be included in the generated image(s) by setting that option to "Y"
rather than "M", or (not recommended on small flash devices) use gpkg
to install the meta-package and the intended dependency packages.

The simple menu structure has been changed to a menuconfig structure
to make it easier to disable the profile package options, which will
be meore helpful if additional package options (language
traslations, themes and optional packages) get added to this config
menu.