Resolve Kconfig recursive dependency error involving kmod-ipt-core
The recursive dependency error reported for kmod-ipt-core stops
"make menuconfig" from displaying and preserving the state of the
profile meta-packages in the output .config.  The simplest and least
invasive way I've found to rectify the problem is to
remove gargoyle-firewall-util as a direct dependency of gargoyle
but still force it's selection whenever gargoyle is selected via the
config mechanism in the gargoyle package Makefile.

(cherry picked from commit c973d6dee70dfa675789420bdb1a11cde7da1275)