plugin-gargoyle-usb-storage-full: add SMB server menu to set dependency
"make menuconfig" can't resolve the "PROVIDES" name as a dependency so
this package gets dropped from the Gargoyle Web Interface sub-menu.

Use a config menu to select which SMB server to as the package dependency
and set the dependencies based on the config variables.

If the unselected SMB server package should be set "=m" that needs to
be addressed separately.

The "PROVIDES" name in both SMB server packages is still useful for
supporting changing SMB server packages in installed systems.  Having this
also allows a user to over-ride the selected SMB server package in a target
config device packages list by specifying '-<unwanted-pkg> <wanted-pkg>').

The "CONFLICTS" setting in both SMB server packages appears to be the
cause of the recursive dependency error reported by "make defconfig" and
"make menuconfig" however it appears not to affect package selection or
display so left unchanged.