Fixed displaying DNS addresses
Removing exactly the IP address, not a string of characters starting with that address.
If the DNS address is in the same address class as LAN, it is removed from the list but should stay.

For example
```
echo "192.168.1.100" | grep -v '192.168.1.1'
(return nothing)

echo "192.168.1.100" | grep -v '192.168.1.1$'
192.168.1.100
```
Also please add these changes to other branch.