ath79 config: optimise package state to suit -basic profile
To maximise the available free space on -basic profile devices, any package
not required for that profile should be set "=m" so they're built for -usb
and -large profile inclusion but not included in -basic images.

Additionally several packages which aren't declared as dependencies
for gargoyle or it's required packages are removed.