ipq806x: convert to use profile meta-packages, add Netgear XR500
The radio firmware packages are pulled in to the respective images
by the device recipes so don't need to be explicitly specified in device
packages.

Instead specify the appropriate profile meta-package: -large for all
IPQ806x devices.

The Netgear XR500 was added for OpenWrt 22.03; it's effectively a repackaged
R7800 and has been widely available.