ath79.default: adjust device profiles, minor cleanups
Add in the lightened USB storage package and With the lightening of the
-usb profile, adjust profiles of 8MB flash devices with USB ports;
some other changes to the -large profile were picked up in review.

Set more packages to =m to minimise unnecessary packages being included
on -basic profile devices; -usb and -large profile devices that need them
will include them as dependencies.