multiple targets: add extroot USB storage package to config
If the extroot USB storage package isn't included in the configs that
build the noshare USB storage package, builds will fail because of the
missing dependency (the custom target's defconfig/menuconfig should
rectify that but update the custom target to match all the others).

swap-utils should be unnecessary given the Busybox swap applets are built
but retained as an installable package for now (some x86 users might want
swaplabel which Busybox doesn't provide).