kmod-usb2: not required as dependency for plugin-gargoyle-usb-storage-noshare
Not all USB2 or USB3 devices declare kmod-usb2 (or kmod-usb3 for that matter)
as device packages; only devices that have drivers that don't directly provide
the usb2/usb3 support require this.  bcm27xx for example is fully USB2 (and USB3
for some devices) capable without kmod-usb2 (or kmod-usb3) installed because the
dwc2 and dwc-otg USB drivers provide the necessary support directly. mvebu, ipq40xx
and ipq806x are other examples of targets where kmod-usb2 and kmod-usb3 aren't
required (ipq806x has a couple of devices with different USB hardware that do
require them but they're the exception).  Targets that require these kmods declare
them as device packages in the image recipes.

It is sufficient that only direct package dependencies be declared for USB class
drivers such as usb-storage, usb-net and usb-serial.