Gargoyle "extras" package: USB ethernet drivers
This meta-package structure is intended as a vehicle for conveniently
adding support for collections of devices or functionality that is
at the edge of Gargoyle's normal requirements; such meta-packages
simplify adding the collections to new configs, custom builds and
standard builds by selecting just the meta-package in a configuration
or for installation on a device.

The initial package groups a number of relatively common USB ethernet
drivers.  These are particularly useful on a range of devices that
only have one normal ethernet port and no switch but one or more USB
ports, such as bcm27xx.  The selected drivers cover devices from
Acquantia, Asix and Realtek.