new target: bcm27xx (RaspberryPi Pi 0, Pi B, Pi 2, Pi 3 and Pi 4)
Add support for all RaspberryPi variants supported by OpenWrt.

All variants get the extra USB ethernet drivers included in images
as this is the most effective way to add a separate ethernet port
to allow functioning as a router; the board configuration also sets
an attached second ethernet port as the WAN (the inbuilt port will
always enumerate as the first ethernet port).  The Pi 4 images also
have support for Realtek 8169 PCI ethernet interfaces as DFRobot
make a "router board" compatible with Pi CM4 modules that includes
this chipset as well as the standard Pi4 ethernet interface.

The squashfs images are probably to be preferred on most SD card
setups and where the power supply to the device is at risk of
disconnection or being powered off rather than shutdown, as issues
with ext4 systems needing to be fsck'ed are commonly reported.
Systems with better performing USB storage and more reliable power
supply can take advantage of ext4 images.