ath79.default: disable devices with unbuildable factory images
Disabled devices:
- D-Link DIR-825 B1
- Trendnet TEW-673gru

Both devices have a 6144kB limit on the size of factory images which
is exceeded by the basic Gargoyle build.  The sysupgrade images are
buildable but the factory images not building is starting to cause
regular build failures despite measures to mitigate this; disabling
these devices avoids this.

The only way these devices could continue to be supported would be to
disable building the factory images therefore requiring an OpenWrt factory
image be initially installed on the device before sysupgrading to Gargoyle.
Doing this would require patching the OpenWrt image recipes which is not
considered worth the effort at this time - if users of these devices come
forward this could be re-visited.