ensure target default packages are included in images
The OpenWrt DEFAULT_PACKAGES mechanism has been found to not reliably
ensure that default packages are included in images unless explicitly
removed via a DEVICE_PACKAGES reference or being superceded by an
equivalent package being installed (e.g. wpad-openssl installed in
place of the default wpad-basic-wolfssl).  The settings in this series
of config patches ensures that packages that should be in images but
aren't defined elsewhere than DEFAULT_PACKAGES are set explicitly to
be included.