ipq806x: pin ethernet interface interrupts to specific CPUs
By default CPU 0 ends up handling all ethernet interface interrupts
even though the default CPU mask for these interrupts should permit
both CPUs to handle them.

Pinning these interrupts to different CPUs noticeably increases the maximum
WAN-to-LAN and LAN-to-WAN bandwidth, especially when QOS is enabled when
throughput can increase by up to a third.  Although most ipq806x devices
seem to now have the same IRQ numbers for these interfaces this hasn't
always been the case and may not be the case for devices I've not been
able to check, so the IRQ numbers to be pinned are dynamically discovered
rather than set as constants.

The IRQs on the wifi interfaces appear to be immutable so the same
strategy can't be applied to those interrupts.